📂 refactor: screens/Recording/Search 폴더 containers 폴더로 이동
