+++ components/GoogleSearchBar/GoogleSearchBar.style.js
@@ -14,3 +14,34 @@ export const Wrapper = styled.View`
 
   z-index: 999;
 `;
+
+export const IconWrapper = styled.View`
+  height: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+`;
+
+export const GooglePlacesAutocompleteStyles = {
+  container: {
+    backgroundColor: globals.colors.GREY_LIGHT,
+    height: 40,
+    borderRadius: 10,
+  },
+  textInputContainer: {
+    height: 40,
+    borderRadius: 10,
+  },
+  textInput: {
+    backgroundColor: globals.colors.GREY_LIGHT,
+    height: 40,
+  },
+  listView: {
+    position: 'absolute',
+    top: 50,
+  },
+  row: { borderBottomWidth: 1, borderColor: globals.colors.GREY_LIGHT },
+  poweredContainer: {
+    display: 'none',
+  },
+};
