+++ components/GoogleSearchBar/index.js
@@ -3,7 +3,7 @@ import { Pressable } from 'react-native';
 import { useNavigation } from '@react-navigation/native';
 import { GooglePlacesAutocomplete } from 'react-native-google-places-autocomplete';
 import getEnvVars from '@/environment';
-import { Wrapper } from './GoogleSearchBar.style';
+import { GooglePlacesAutocompleteStyles, IconWrapper, Wrapper } from './GoogleSearchBar.style';
 import { globals } from '@styles/globals';
 import BackBtn from '@assets/images/back.svg';
 import CloseBtn from '@assets/images/close.svg';
@@ -28,44 +28,29 @@ const GoogleSearchBar = () => {
         }}
         onPress={(data, details) => console.log(details.geometry?.location)}
         renderLeftButton={() => (
-          <SearchBtn
-            width={20}
-            height={20}
-            fill={globals.colors.GREY_DEF_DARK}
-            style={{ marginLeft: 10, alignSelf: 'center', position: 'relative', top: -2 }}
-          />
+          <IconWrapper>
+            <SearchBtn
+              width={20}
+              height={20}
+              fill={globals.colors.GREY_DEF_DARK}
+              style={{ marginLeft: 10 }}
+            />
+          </IconWrapper>
         )}
         renderRightButton={() => (
-          <Pressable onPress={() => this.textInput.clear()}>
-            <CloseBtn
-              width={15}
-              height={15}
-              fill={globals.colors.GREY_DEF_DARK}
-              style={{ marginRight: 10, alignSelf: 'center', position: 'relative', top: 17 }}
-            />
-          </Pressable>
+          <IconWrapper>
+            <Pressable onPress={() => this.textInput.clear()}>
+              <CloseBtn
+                width={15}
+                height={15}
+                fill={globals.colors.GREY_DEF_DARK}
+                style={{ marginRight: 10 }}
+              />
+            </Pressable>
+          </IconWrapper>
         )}
         textInputProps={{ placeholderTextColor: '#A5A5A5' }}
-        styles={{
-          container: {
-            backgroundColor: globals.colors.GREY_LIGHT,
-            borderRadius: 10,
-          },
-          textInputContainer: {
-            borderRadius: 10,
-          },
-          textInput: {
-            backgroundColor: globals.colors.GREY_LIGHT,
-          },
-          listView: {
-            position: 'absolute',
-            top: 50,
-          },
-          row: { borderBottomWidth: 1, borderColor: globals.colors.GREY_LIGHT },
-          poweredContainer: {
-            display: 'none',
-          },
-        }}
+        styles={GooglePlacesAutocompleteStyles}
       />
       <Pressable onPress={() => navigation.navigate('Home')}>
         <BackBtn style={{ marginLeft: 10 }} />
