+++ screens/Recording/Recording.jsx
@@ -1,5 +1,5 @@
 import React, { useRef, useState } from 'react';
-import Search from './Search';
+import Search from '@containers/Recording/Search';
 import Header from '@components/Header';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
 import RouteRunning from './RouteRunning';
